+++ bodhi/server/consumers/composer.py
@@ -41,6 +41,8 @@ from urllib.request import urlopen
 
 import jinja2
 import fedora_messaging
+from twisted.internet import reactor
+from twisted.internet.threads import blockingCallFromThread
 
 from bodhi.messages.schemas import compose as compose_schemas, update as update_schemas
 from bodhi.server import bugs, initialize_db, buildsys, notifications, mail
@@ -187,8 +189,13 @@ class ComposerHandler(object):
         message = message.body
         resume = message.get('resume', False)
         agent = message.get('agent')
-        notifications.publish(compose_schemas.ComposeStartV1.from_dict(dict(agent=agent)),
-                              force=True)
+        # This callback is run in a thread by Twisted, and must therefore call
+        # the messaging API with the proper Twisted wrapper to ensure thread
+        # safety.
+        blockingCallFromThread(
+            reactor, notifications.publish,
+            compose_schemas.ComposeStartV1.from_dict(dict(agent=agent)),
+            force=True)
 
         results = []
         threads = []
@@ -362,11 +369,14 @@ class ComposerThread(threading.Thread):
 
         log.info('Running ComposerThread(%s)' % self.id)
 
-        notifications.publish(compose_schemas.ComposeComposingV1.from_dict(
-            dict(repo=self.id,
-                 updates=[' '.join([b.nvr for b in u.builds]) for u in self.compose.updates],
-                 agent=self.agent,
-                 ctype=self.ctype.value)),
+        blockingCallFromThread(
+            reactor,
+            notifications.publish,
+            compose_schemas.ComposeComposingV1.from_dict(
+                dict(repo=self.id,
+                     updates=[' '.join([b.nvr for b in u.builds]) for u in self.compose.updates],
+                     agent=self.agent,
+                     ctype=self.ctype.value)),
             force=True,
         )
 
@@ -477,7 +487,9 @@ class ComposerThread(threading.Thread):
             update.remove_tag(update.release.pending_testing_tag,
                               koji=buildsys.get_session())
         update.request = None
-        notifications.publish(
+        blockingCallFromThread(
+            reactor,
+            notifications.publish,
             update_schemas.UpdateEjectV1.from_dict(
                 dict(
                     repo=self.id,
@@ -523,8 +535,11 @@ class ComposerThread(threading.Thread):
             success (bool): True if the compose had been successful, False otherwise.
         """
         log.info('Thread(%s) finished.  Success: %r' % (self.id, success))
-        notifications.publish(compose_schemas.ComposeCompleteV1.from_dict(dict(
-            dict(success=success, repo=self.id, agent=self.agent, ctype=self.ctype.value))),
+        blockingCallFromThread(
+            reactor,
+            notifications.publish,
+            compose_schemas.ComposeCompleteV1.from_dict(dict(
+                dict(success=success, repo=self.id, agent=self.agent, ctype=self.ctype.value))),
             force=True,
         )
 
@@ -692,7 +707,7 @@ class ComposerThread(threading.Thread):
                 UpdateRequest.testing: update_schemas.UpdateCompleteTestingV1
             }
             message = messages[update.request].from_dict(dict(update=update, agent=agent))
-            notifications.publish(message, force=True)
+            blockingCallFromThread(reactor, notifications.publish, message, force=True)
 
     @checkpoint
     def modify_bugs(self):
@@ -1228,8 +1243,11 @@ class PungiComposerThread(ComposerThread):
         """Wait for a repo signature to appear."""
         # This message indicates to consumers that the repos are fully created and ready to be
         # signed or otherwise processed.
-        notifications.publish(compose_schemas.RepoDoneV1.from_dict(
-            dict(repo=self.id, agent=self.agent, path=self.path)),
+        blockingCallFromThread(
+            reactor,
+            notifications.publish,
+            compose_schemas.RepoDoneV1.from_dict(
+                dict(repo=self.id, agent=self.agent, path=self.path)),
             force=True)
         if config.get('wait_for_repo_sig'):
             self.save_state(ComposeState.signing_repo)
@@ -1266,8 +1284,11 @@ class PungiComposerThread(ComposerThread):
             Exception: If no folder other than "source" was found in the compose_path.
         """
         log.info('Waiting for updates to hit the master mirror')
-        notifications.publish(compose_schemas.ComposeSyncWaitV1.from_dict(
-            dict(repo=self.id, agent=self.agent)),
+        blockingCallFromThread(
+            reactor,
+            notifications.publish,
+            compose_schemas.ComposeSyncWaitV1.from_dict(
+                dict(repo=self.id, agent=self.agent)),
             force=True)
         compose_path = os.path.join(self.path, 'compose', 'Everything')
         checkarch = None
@@ -1301,8 +1322,11 @@ class PungiComposerThread(ComposerThread):
                 continue
             if newsum == checksum:
                 log.info("master repomd.xml matches!")
-                notifications.publish(compose_schemas.ComposeSyncDoneV1.from_dict(
-                    dict(repo=self.id, agent=self.agent)),
+                blockingCallFromThread(
+                    reactor,
+                    notifications.publish,
+                    compose_schemas.ComposeSyncDoneV1.from_dict(
+                        dict(repo=self.id, agent=self.agent)),
                     force=True)
                 return